Fixes fifo buffer depth
Changes fifo buffer design so that it can have the right depth and
correctly keep its counter consistent (differentiate the full and empty
states).
With that it is possible to receive data when it's full given
that, at the same time, data is being pulled out. That explains the change at
credit_o signal of thor_buffer.

Commit: Fix, Mod

TaskNumber: #63